small fix in content creator

+++ demisto_sdk/commands/create_artifacts/content_creator.py
@@ -307,8 +307,9 @@ class ContentCreator:
             if os.path.isdir(path):
                 non_circle_tests = glob.glob(os.path.join(path, '*'))
                 for new_path in non_circle_tests:
-                    print(f'copying path {new_path}')
-                    shutil.copyfile(new_path, os.path.join(self.test_bundle, os.path.basename(new_path)))
+                    if os.path.isfile(new_path):
+                        print(f'copying path {new_path}')
+                        shutil.copyfile(new_path, os.path.join(self.test_bundle, os.path.basename(new_path)))
 
             else:
                 # test playbooks in test_playbooks_dir in packs can start without playbook* prefix
